# pottershouse/pottershouse-backend/openapi.yaml
openapi: 3.0.3
info:
  title: Potters House API
  version: 1.0.0
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorWrapper:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    nullable: true
                  message:
                    type: string
            conflictingBookings:
              type: array
              items: {}
            suggested_alternatives:
              type: array
              items: {}
...
    Booking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_name:
          type: string
        phone:
          type: string
        email:
          type: string
        event_type:
          type: string
        event_date_start:
          type: string
          format: date-time
        event_date_end:
          type: string
          format: date-time
        event_all_day:
          type: boolean
        guests:
          type: integer
        package_id:
          type: string
          format: uuid
          nullable: true
        dietary_needs:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
        workflow_status:
          type: string
        deposit_paid:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    BookingAdminUpdateRequest:
      type: object
      properties:
        customer_name:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        event_type:
          type: string
        event_date_start:
          type: string
          format: date-time
        event_date_end:
          type: string
          format: date-time
        event_all_day:
          type: boolean
        guests:
          type: integer
        package_id:
          type: string
          format: uuid
          nullable: true
        dietary_needs:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
        workflow_status:
          type: string
        deposit_paid:
          type: boolean
...
paths:
...
  /admin/bookings/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get booking (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
    put:
      summary: Update booking (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingAdminUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
    patch:
      summary: Update booking (admin)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingAdminUpdateRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorWrapper'
...